requirements.txt
matplotlib==3.9.4
networkx==3.4.2
qutip==5.0.4
Flask==3.1.0
flask-cors==5.0.1
Flask-SocketIO==5.5.1
requests==2.32.3
waitress==3.0.2
langchain>=0.1.0
langchain-groq>=0.1.0
python-dotenv>=1.0.0
regex>= 2024.11.6
